Fix some compilation warnings

+++ lib/cpp/examples/nlateral_teleop.cc
@@ -187,7 +187,7 @@ void DisplayUsage() {
   std::cout << "    fXX.X - scale force by this positive float\n";
 
   auto pad = [](const std::string str, int size) {
-    if (str.size() < size) {
+    if (static_cast<int>(str.size()) < size) {
       return str + std::string(size - str.size(), ' ');
     }
     return str;
